+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Copy project dashboards when a project is copied via `Project#copy`
 - Fix `NoMethodError` in `DashboardContentProject` `projectinformation` block when adding blocks to a system-default project dashboard outside any project context
 - Fix PostgreSQL error in `Project#assignable_principals` and `Principal.assignable_for_issues` when chained with `.sorted`
+- Add option to include subproject news in the dashboard news block #164
 - d3plus updated to 3.1.6
 
 ## 4.4.0